remove unecessary toasts

+++ frontend/src/contexts/CartContext.jsx
@@ -10,6 +10,7 @@ export function CartProvider({ children }) {
   const { user } = useContext(AuthContext);
   const [cart, setCart] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
+  const [isSubmitting, setIsSubmitting] = useState();
 
   useEffect(() => {
     const getCart = async () => {
@@ -29,6 +30,8 @@ export function CartProvider({ children }) {
   }, [user?.username]);
 
   const addToCart = async (productId, quantity) => {
+    if (isSubmitting) return;
+    setIsSubmitting(true);
     try {
       const newCartItem = await shoppingCartService.addCartItem(productId, quantity);
       const { id, name, pictureUrl, price } = newCartItem.product;
@@ -43,22 +46,26 @@ export function CartProvider({ children }) {
 
         return [...cartItems, { id, name, pictureUrl, price, productId, quantity }];
       });
-      toast.success('Product(s) added to cart succesfully.');
     } catch (error) {
-      toast.error('Could not add product to cart.');
+      toast.error(error.response.data.error);
+    } finally {
+      setIsSubmitting(false);
     }
   };
 
   const removeFromCart = async (productId) => {
+    if (isSubmitting) return;
+    setIsSubmitting(true);
     try {
       await shoppingCartService.removeCartItem(productId);
 
       const newCart = await shoppingCartService.getShoppingCartByUserId();
 
       setCart(newCart);
-      toast.success('Product(s) deleted from cart succesfully.');
     } catch (error) {
       toast.error('Could not delete product from cart.');
+    } finally {
+      setIsSubmitting(false);
     }
   };
